% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators_analysis.R
\name{lcl_interim}
\alias{lcl_interim}
\title{One-sided CE-consistent lower confidence limit (early stop; surrogate scale)}
\usage{
lcl_interim(est1, se1, b_ref)
}
\arguments{
\item{est1}{Interim surrogate-projected estimate.}

\item{se1}{Interim standard error.}

\item{b_ref}{Reference boundary.}
}
\value{
Numeric scalar.
}
\description{
One-sided CE-consistent lower confidence limit (early stop; surrogate scale)
}
