% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{popshiftce_defaults}
\alias{popshiftce_defaults}
\title{Default parameter template for PopShiftCE examples}
\usage{
popshiftce_defaults()
}
\value{
A named list.
}
\description{
Returns a named list of baseline parameters similar to the paper's baseline
setting. Users can modify selected entries and pass them into
\code{build_ce_lookup()} / \code{simulate_trial_ce()} / \code{simulate_trials_ce()}.
}
