% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_stage1_and_T1}
\alias{gen_stage1_and_T1}
\title{Stage-1 surrogate-only interim statistic T1}
\usage{
gen_stage1_and_T1(n1, muX_C, muX_T, sigmaX, gamma0, gamma1)
}
\arguments{
\item{n1}{per-arm Stage-1 size}

\item{muX_C}{control-arm mean of X}

\item{muX_T}{treatment-arm mean of X}

\item{sigmaX}{sd of X}

\item{gamma0, gamma1}{external calibration coefficients for Yhat = gamma0 + gamma1 * X}
}
\value{
list(T1, est1, se1)
}
\description{
Stage-1 surrogate-only interim statistic T1
}
\keyword{internal}
