% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_stage2_new}
\alias{gen_stage2_new}
\title{Stage-2 new participants under partial population shift}
\usage{
gen_stage2_new(
  n2,
  mu0,
  delta,
  theta,
  eta,
  sigmaY,
  piZ,
  error_type = c("normal", "t", "skew")
)
}
\arguments{
\item{error_type}{one of "normal","t","skew"}
}
\value{
data.frame with columns stage, T, Z, Y, is_new
}
\description{
Stage-2 new participants under partial population shift
}
\keyword{internal}
