% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lcl_final}
\alias{lcl_final}
\title{One-sided CE-consistent LCL (final stage)}
\usage{
lcl_final(Delta_hat, se, c_t1)
}
\arguments{
\item{Delta_hat}{numeric}

\item{se}{numeric}

\item{c_t1}{conditional critical value c(t1)}
}
\value{
numeric
}
\description{
One-sided CE-consistent LCL (final stage)
}
