% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lcl_interim}
\alias{lcl_interim}
\title{One-sided CE-consistent LCL (early stop, surrogate scale)}
\usage{
lcl_interim(est1, se1, b_ref)
}
\arguments{
\item{est1}{interim estimate (surrogate-projected)}

\item{se1}{interim se}

\item{b_ref}{reference boundary}
}
\value{
numeric
}
\description{
One-sided CE-consistent LCL (early stop, surrogate scale)
}
