% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h0_ce.R
\name{one_H0_pair}
\alias{one_H0_pair}
\title{One H0 pair (T1, S_final) for CE calibration (design-fixed target)}
\usage{
one_H0_pair(
  n1,
  n2,
  muX_C,
  sigmaX,
  gamma0,
  gamma1,
  mu0,
  sigmaY,
  theta,
  eta,
  piZ,
  pi_fixed = 0.5,
  error_type = c("normal", "t", "skew"),
  rho_XY = 0.3
)
}
\description{
One H0 pair (T1, S_final) for CE calibration (design-fixed target)
}
\keyword{internal}
