% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ce_mapping}
\alias{plot_ce_mapping}
\title{Plot CE mapping (diagnostics): joint (T1, S_final), e(t1), c(t1)}
\usage{
plot_ce_mapping(H0_dt, lookup)
}
\arguments{
\item{H0_dt}{data.frame with columns z1, zf (H0 pairs)}

\item{lookup}{object from \code{\link[=build_ce_lookup]{build_ce_lookup()}}}
}
\value{
a named list of ggplot objects: pA, pB, pC
}
\description{
Plot CE mapping (diagnostics): joint (T1, S_final), e(t1), c(t1)
}
